requirements.txt
djangorestframework-simplejwt
pytest-playwright
playwright
pyjwt
pytest-xdist
pytest-asyncio
allure-pytest
pytest-html
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
greenlet==3.2.4
h11==0.16.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
playwright==1.56.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
text-unidecode==1.3
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.3.2
